requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-appconf==1.0.2
django-avatar==4.1.0
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-filter==2.1.0
django-friendship==1.8.0
django-heroku==0.3.1
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
oauthlib==3.0.1
Pillow==5.4.1
psycopg2==2.7.7
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
whitenoise==4.1.2
